refactor(mentor): BestProducts와 AllProducts 분리
